verify, merkle, rekor: move `verify_checkpoint`
